requirements.txt
boto3
numpy
ruamel.yaml
scikit-learn
scipy
six
tqdm
watchtower
